Create motion step card component
