+++ app/page.tsx
@@ -1,5 +1,3 @@
 import { LandingPage } from "@/features/LandingPage";
 
-export default function Page() {
-  return <LandingPage />;
-}
+export default LandingPage;
