+++ features/AnalyzePage/components/MotionStepCard/index.ts
@@ -1,0 +1,1 @@
+export { default as MotionStepCard } from "./MotionStepCard.component";
