+++ features/AnalyzePage/index.ts
@@ -1,0 +1,1 @@
+export { default as AnalyzePage } from "./AnalyzePage.component";
