+++ features/index.ts
@@ -1,1 +1,3 @@
+export * from "./AnalyzePage";
 export * from "./InstagramDashboard";
+export * from "./LandingPage";
